requirements.txt
opencv-python
matplotlib
scikit-image
filterpy
lap
scipy
ultralytics
roboflow
ipykernel
fastapi
uvicorn
mysql-connector-python
schedule
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126